feat: show a preview of the charts in the chart overview
