+++ frontend/src/routes/(app)/project/[uuid]/[name]/chart/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
-	import ChartHomeBlock from '$lib/components/chart/ChartHomeBlock.svelte';
+	import ChartCard from '$lib/components/chart/ChartCard.svelte';
 	import { charts, project } from '$lib/stores';
 	import { chartDefaults } from '$lib/util/charts';
 	import { ChartType, ZenoService } from '$lib/zenoapi';
@@ -41,9 +41,9 @@
 		{/if}
 	</div>
 	<div class="mb-6 h-0.5 w-full rounded-full bg-grey-light" />
-	<div class="flex flex-wrap overflow-y-auto">
+	<div class="mb-4 grid h-full grid-cols-home content-start gap-5 overflow-y-auto">
 		{#each $charts.sort((a, b) => a.id - b.id) as chart}
-			<ChartHomeBlock {chart} />
+			<ChartCard {chart} />
 		{/each}
 	</div>
 </div>
